+++ src/Console/CleanOutdatedReleaseData.php
@@ -39,7 +39,8 @@ class CleanOutdatedReleaseData extends Command
 
         $releases = $releaseModel::query()
             ->where('created_at', '<=', Carbon::now()->subDays($days))
-            ->whereNot('id', $releaseModel::getLastRelease()?->id)
+            ->where('updated_at', '<=', Carbon::now()->subDays(7))
+            ->whereNot('id', $releaseModel::getActiveRelease()?->id)
             ->get();
 
         try {